pyproject.toml
[project]
name = "ai-invoice-auditor"
version = "0.1.0"
description = "Agentic AI Invoice Auditor"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "langgraph>=0.0.20",
    "langchain>=0.1.0",
    "langchain-cohere>=0.1.0",
    "langchain-community>=0.0.10",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "httpx>=0.27.0",
    "qdrant-client>=1.8.0",
    "pyyaml>=6.0.1",
    "litellm>=1.30.0",
    "pymupdf>=1.24.0",
    "instructor>=1.0.0",
    "fastmcp>=0.1.0",
    "langchain-text-splitters>=0.0.1",
    "langdetect>=1.0.9",
    "langfuse>=2.0.0",
    "rich>=13.7.0",
    "fpdf2>=2.7.0",
    "boto3>=1.42.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# --- FIX: Explicitly tell hatch to include 'src' ---
[tool.hatch.build.targets.wheel]
packages = ["src"]
